store: Move creation of HistoryEvent into history_event module
